Add files via upload
This is the code that is used to backtest in which the 5 EMA stretegy is work to analysis the Stock Market data. first we have to provide the OHLC data that is the candle stick data  it will apply the 5 EMA on the OHLC data that is open high low close data. after we provide the ohlc data the output is it will generate the csv file and this file contain the backtesting result